Develop methods AddNewAllocation and ModifyExistingAllocation
Introduced two separated allocation lists:
m_addtsAllocationList stores all the allocations granted by the AP through an ADDTS request.
m_globalAllocationList is the one sent in an DMG beacon frame. It adds to  m_addtsAllocationList a series of broadcast CBAPs where there is DTI time not allocated. The function which performing this feature is going to be implemented in later commits