Calling StartBeaconInterval from dmgWifiScheduler
The function calculateBTIvariables of DmgApWifiMac is scheduled when the BI ends. Together with the evaluation of ADDTS request at DmgWifiScheduler. It happens that the former method is called before evaluating the requests (tough at the same simulation time) resulting in an empy extended schedule element that impacts the right calculation of the DMG beacon duration.