Add DmgWifiScheduler class which handles the scheduling features for 80211ad (#10)
* Create class DmgWifiScheduler

* Add custom dmg error model to lroom scenario

* Add interface between DmgApWifiMac and DmgWifiScheduler

* Setup callback to handle a received ADDTS request at DmgScheduler

* Connect BI start and DTI start traces to DmgWifiScheduler

* Move methods of type Add Allocation from DmgApWifiMac to DmgScheduler

* Move old evaluate scripts in examples/wireless/80211ad

Examples are disabled by default in ns3-80211ad due to old compatibility issue. Therefore, we should not have any problem.

* Create callback to handle a DELTS request at DmgWifiScheduler

* Add methods to store and manage the ADDTS requests received

* Add map to store the allocated ADDTS requests

m_allocatedAddtsRequests stores the ADDTS requests that have been accepted and allocated by the PCP/AP.
The original allocation parameters for each request  are stored in this map.

* Add two methods to handle an ADDTS request

The ADDTS request is handled differently whether or not the allocation already exists. 
ModifyExistingAllocation handles the acceptance or rejection when the AP receives an ADDTS request related to an existing allocation. Whereas, AddNewAllocation handles the acceptance or rejection  of an ADDTS request NOT related to an existing one.

* Add function  to get the allocation duration (between min and max))

* Develop methods AddNewAllocation and ModifyExistingAllocation

* Create method to place broadcast CBAPs among requested allocations

* Modify AddBroadcastCbapAllocations based on the interAllocation distance

* Add guard time between adjacent allocations

* Moved CalculationBTIVariables at the start of beacon interval

The BHI duration at DmgWifiScheduler differed from the one at DmgAp, because the above method was called at the start of the BTI. This method calculates the duration of the BTI based on duration of a DMG beacon.

* Temporary Fix to issue 9. Create dummy ESE with MAX size

* Add Scheduler creation from the DmgWifiHelper

* Evaluate an ADDTS request as soon as it is received

* Send ADDTS response to destination STA in any case

According to the standard, an ADDTS response to the destination STA of the allocation is sent even if its status code is different from Success.

* Add BasicDmgWifiScheduler which implements the abstract DmgWifiScheduler

The child class BasicDmgWifiScheduler implements the pure virtual functions declared by DmgWifiScheduler. 
It provides the implementation of a basic admission control policy and DTI scheduling strategy for 80211ad.