Add simple version of AddNewAllocation and introduced UpdateAvailableSlots
This version of AddNewAllocation implements a periodic-only acceptance 
policy. The assumption is that ADDTS requests include only periodic 
traffic. The member variable m_availableSlots has been introduced to 
manage this new type of traffic, as the previous variables were not 
enough. For this reason, a specific method has been created to update 
this variable.