Add AdjustExistingAllocation with proper documentation
As of now the input parameters passed to this method are useless. Being m_availableSlots of key importance to manage the allocations, this method simply update the list of available slots. This is a consequence of the fact that AdjustExistingAllocations is triggered after a DELTS request or after a clean up procedure of all the non-pseudostatic allocations.