Add support to periodic allocations  (#20)
* Add simple version of AddNewAllocation and introduced UpdateAvailableSlots

This version of AddNewAllocation implements a periodic-only acceptance 
policy. The assumption is that ADDTS requests include only periodic 
traffic. The member variable m_availableSlots has been introduced to 
manage this new type of traffic, as the previous variables were not 
enough. For this reason, a specific method has been created to update 
this variable.

* Create draft version of AddBroadcastCbapAllocations

In this version, the method fills the remaining available slots with 
broadcast CBAPs.

* Simplify AddBroadcastCbapAllocations

Now the Cbap Allocations are added based on m_allocationLists and 
m_remainingTime

* Fixed spStart update to follow standard guidelines

* Simplified AddBroadcastCbapAllocations

In this version of the method we simply fill each available time slot in 
the DTI with a CBAP. At the end, we order the m_allocationList for 
possible further future improvements.

* Update doxy for UpdateAvailableSlots

* Update AddNewAllocation to include more than one periodic allocation.

This version of AddNewAllocation allows the scheduling of more than one periodic allocation in the DTI. Also, non-periodic allocations can be allocated either before or after the periodic ones.

* Optimized AddBroadcastCbapAllocation with range-based for loops

* Add comments to AddNewAllocation and fix indentation

* Modify UpdateAvailableSlots to consider changes to existing allocations

* Modify UpdateStartAndRemainingTime to keep it consistent

Considering how we manage the periodic requests, this method is not of key importance. However, to keep its consistency we update m_remainingDtiTime and we initialize m_availableSlots if no allocations have been requested yet.

* Add AdjustExistingAllocation with proper documentation

As of now the input parameters passed to this method are useless. Being m_availableSlots of key importance to manage the allocations, this method simply update the list of available slots. This is a consequence of the fact that AdjustExistingAllocations is triggered after a DELTS request or after a clean up procedure of all the non-pseudostatic allocations.

* Removed m_minBroadcastCbapDuration and m_interAllocationDistance

* Fix bug in AdjustExistingAllocations 

The previous version of the method didn't consider the case where one or more periodic allocations were left in m_addtsAllocationList. Now this condition is checked and dealt with accordingly.

* Overload of UpdateAvailableSlots

Now the method is divided into two distinct versions that address two different conditions, i.e., new allocation added to the list of allocations and a change in previously allocated slots.

* Update consistently m_remainingDtiTime and avoid redundant code

* Fix GetAvailableBlocks and improve doxy

* Modify GetAvailableBlocks and avoid redundant code in AddNewAllocation

GetAvailableBlocks has been modified to return a list containing the starting time of each block that can be allocated. Accordingly, AddNewAllocation has been modified to consider a non-periodic allocation has an edge case of the periodic one.

* Optimize UpdateAvailableSlot

m_availableSlots has been changed from vector to list and both versions of UpdateAvailableSlots have been updated to perform in-place insertion/deletion.